chore: refactor `JwtAuthenticationResponse` package for consistency across modules
Moved `JwtAuthenticationResponse` from `keycloak.response` to `response.jwt`. Updated all imports and references in `lsadf_core`, `lsadf_api`, and `lsadf_admin` modules accordingly.
