+++ lsadf_api/src/test/java/com/lsadf/application/bdd/when/BddCharacteristicsWhenStepDefinitions.java
@@ -24,10 +24,10 @@ import com.lsadf.application.bdd.BddUtils;
 import com.lsadf.application.controller.constant.ApiPathConstants;
 import com.lsadf.application.controller.game.characteristics.CharacteristicsController;
 import com.lsadf.core.domain.game.characteristics.Characteristics;
-import com.lsadf.core.infra.web.client.keycloak.response.JwtAuthenticationResponse;
 import com.lsadf.core.infra.web.request.game.characteristics.CharacteristicsRequest;
 import com.lsadf.core.infra.web.response.ApiResponse;
 import com.lsadf.core.infra.web.response.game.characteristics.CharacteristicsResponse;
+import com.lsadf.core.infra.web.response.jwt.JwtAuthenticationResponse;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.When;
 import lombok.extern.slf4j.Slf4j;
