+++ lsadf_api/src/test/java/com/lsadf/application/bdd/when/BddCurrencyWhenStepDefinitions.java
@@ -24,10 +24,10 @@ import com.lsadf.application.bdd.BddUtils;
 import com.lsadf.application.controller.constant.ApiPathConstants;
 import com.lsadf.application.controller.game.currency.CurrencyController;
 import com.lsadf.core.domain.game.currency.Currency;
-import com.lsadf.core.infra.web.client.keycloak.response.JwtAuthenticationResponse;
 import com.lsadf.core.infra.web.request.game.currency.CurrencyRequest;
 import com.lsadf.core.infra.web.response.ApiResponse;
 import com.lsadf.core.infra.web.response.game.currency.CurrencyResponse;
+import com.lsadf.core.infra.web.response.jwt.JwtAuthenticationResponse;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.When;
 import lombok.extern.slf4j.Slf4j;
